funciona compra desde cardDescrition
