--- frontend/src/components/CardAmountConter.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-
-function CardAmountConter(){
-
-    return(
-        <>
-        <p className="counter-button">-</p>
-        <p className="counter-button">+</p>
-        </>
-    )
-    
-}
